Add pytest test suite for versatility module
- Cherry-pick utils/io, utils/parsing and their tests from dev branch
- Extend conftest.py with MuxVizRunner, toy network (10 nodes, 3 layers),
  supra-adjacency/interaction fixtures, and comparison helpers
- Add test_versatility.py with three test classes:
  - TestVersatilityCorrectness (34 tests): shapes, ranges, invariants
  - TestVersatilityBackendComparison (7 tests): muxvizpy vs hornet backends
  - TestVersatilityReference (13 tests): vs muxViz R via Singularity container
- Fix int64 casting bug in compute_multi_authority_centrality (AAT.data += eps)

+++ src/MuxVizPy/versatility.py
@@ -552,7 +552,7 @@ def compute_multi_authority_centrality(
     np.ndarray
         Max-normalized authority centrality per physical node, shape (n,).
     """
-    AAT = adj.T.dot(adj)
+    AAT = adj.T.dot(adj).astype(np.float64)
     AAT.data += eps
     eigen_search_failed = True
     count_eigen_attempts = 0
